Add: Stat class added and LeagueStats inherits from this class now. Allows easier readability as all CSV reading methods are inherited from Stats.
