+++ lib/league_stats.rb
@@ -1,9 +1,10 @@
 require 'csv'
+require_relative 'stats'
 require_relative 'game'
 require_relative 'team'
 require_relative 'game_teams'
 
-class LeagueStats
+class LeagueStats < Stats
   attr_reader :games, :teams, :game_teams
 
   def initialize(data)
@@ -12,39 +13,39 @@ class LeagueStats
     @game_teams = build_game_teams(data)
   end
 
-  def build_games(data)
-    games = []
-    info = CSV.parse (File.read(data[:games])), headers: true, header_converters: :symbol
-    info.each do |row|
-      games << Game.new(row)
-    end
-    games
-  end
+  # def build_games(data)
+  #   games = []
+  #   info = CSV.parse (File.read(data[:games])), headers: true, header_converters: :symbol
+  #   info.each do |row|
+  #     games << Game.new(row)
+  #   end
+  #   games
+  # end
 
-  def build_teams(data)
-    teams = []
-    team_info = CSV.parse (File.read(data[:teams])), headers: true, header_converters: :symbol
-    team_info.each do |row|
-      teams << Team.new(row)
-    end
-    teams
-  end
+  # def build_teams(data)
+  #   teams = []
+  #   team_info = CSV.parse (File.read(data[:teams])), headers: true, header_converters: :symbol
+  #   team_info.each do |row|
+  #     teams << Team.new(row)
+  #   end
+  #   teams
+  # end
 
-  def build_game_teams(data)
-    game_teams = []
-    game_teams_info = CSV.parse (File.read(data[:game_teams])), headers: true, header_converters: :symbol
-    game_teams_info.each do |row|
-      game_teams << GameTeam.new(row)
-    end
-    game_teams
-  end
+  # def build_game_teams(data)
+  #   game_teams = []
+  #   game_teams_info = CSV.parse (File.read(data[:game_teams])), headers: true, header_converters: :symbol
+  #   game_teams_info.each do |row|
+  #     game_teams << GameTeam.new(row)
+  #   end
+  #   game_teams
+  # end
 
   def count_of_teams
     @teams.count
   end
 
   def best_offense
-    avg_goals = average_team_goals_all_seasons
+    avg_goals = average_team_goals_all_seasons(team_goals_hash, games_per_team)
     best_offense = avg_goals.max_by do |team, avg|
       avg
     end
@@ -53,7 +54,7 @@ class LeagueStats
   end
 
   def worst_offense
-    avg_goals = average_team_goals_all_seasons
+    avg_goals = average_team_goals_all_seasons(team_goals_hash, games_per_team)
     worst_offense = avg_goals.min_by do |team, avg|
       avg
     end
@@ -61,7 +62,7 @@ class LeagueStats
     worst_team.name
   end
 
-  def average_team_goals_all_seasons
+  def average_team_goals_all_seasons(goals_hash, games_hash)
     goals_hash = team_goals_hash
     games_hash = games_per_team
     avg_goals = Hash.new
